content-api/requirements.txt
flask
google-cloud-firestore
gunicorn